commit dev christ

+++ Freads/app/Http/Controllers/ProfileController.php
@@ -9,6 +9,7 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\View\View;
 use App\Models\User;
+use App\Models\Ad;
 
 class ProfileController extends Controller
 {
@@ -60,14 +61,44 @@ class ProfileController extends Controller
     }
 
     public function AdminUser(){
+        if(auth()->user()->admin == 0){
+            return redirect()->route('acceuil');
+        }
         $users = User::paginate(15);
 
         return view('admin.user', compact('users'));
     }
 
     public function deleteUser($id){
+        if(auth()->user()->admin == 0){
+            return redirect()->route('acceuil');
+        }
         $user= User::findOrFail($id);
         $user->delete();
         return redirect()->route('admin.user');
+    }
+    
+    public function editUser($id){
+        if(auth()->user()->admin == 0){
+            return redirect()->route('acceuil');
+        }
+        $user = User::findOrFail($id);
+        $ads = Ad::where("user_id", "=", $user->id)->get();
+        return view('admin.modify', [
+            "user" => $user,
+            "ads" => $ads
+        ]);
+    }
+
+    public function modifyUser(Request $request, $id){
+        if(auth()->user()->admin == 0){
+            return redirect()->route('acceuil');
+        }
+        $user= User::where("id", "=", $id)->update(["admin" => $request->admin]);
+        /*$user->update($request->all());
+        if($request->validate(['admin' => ['required'], 'exists:users,admin'.User::class])){
+            $user->update($request->all());
+        }*/
+       return redirect()->route('admin.user');
     }
 }
